+++ app/src/main/java/com/bodan/maplecalendar/presentation/views/skill/linkskill/LinkSkillInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.bodan.maplecalendar.presentation.views.skill
+package com.bodan.maplecalendar.presentation.views.skill.linkskill
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
