+++ app/src/main/res/layout/fragment_hyper_skill_info.xml
@@ -18,7 +18,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".presentation.views.skill.HyperSkillInfoFragment">
+            tools:context=".presentation.views.skill.hyperskill.HyperSkillInfoFragment">
 
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/rv_hyper_skill_info"
